toolkit/cli/relay/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
fastapi==0.115.5
uvicorn==0.32.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1